feat: Add Trees component to display game data
This commit adds a new component called Trees.vue, which is responsible for rendering a forest of trees based on the game data provided. The component uses the D3 library to create and manipulate SVG elements.

The Trees component receives an array of game data as a prop and uses it to create instances of the Tree class for each game. The Tree class represents an individual tree in the forest and is responsible for calculating the tree's initial length, maximum depth, and branch angle based on the game data.

The init method of the Trees component initializes the forest by creating instances of the Tree class for each game and storing them in the trees array. The draw method then iterates over the trees array and calls the init and draw methods of each tree to render them on the SVG.

This commit also includes some minor changes to other files, such as updating the main.scss file to change the stroke color of the trees and adding a new page called index.vue to display the game data.

The purpose of this commit is to add the functionality to display the game data as a forest of trees, providing a visual representation of the data for better user experience.